requirements.txt
dlib==19.24.0
face-recognition==1.3.0
numpy==1.24.2
opencv-python~=4.8.1.78
pymongo~=4.6.1
python-dotenv~=1.0.0